added jupyterlab 4 compatibility info & versioning change

+++ src/jansfortrankernel/kernel.py
@@ -42,4 +42,5 @@ class jansfortrankernel(Kernel):
                }
 
     def do_shutdown(self, restart):
-        shutil.rmtree(workingdir)
+        if os.path.exists(workingdir):
+            shutil.rmtree(workingdir)
